Simplify achievement selection logic
Simplify handling and interaction with the GUI when an
achievement is "active" - that is, when it is pending modification.

Clean up makefile some

Make makefile be sensitive to header file changes - For now,
make every file sensitive to any header change for simplicity.
This prevents build corruptions on header changes.